✨feat: 입력을 담당하는 핸들러로 변경 구현
